+++ cleo_plugins/MemoryOperations/MemoryOperations.cpp
@@ -220,6 +220,14 @@ public:
         }
 
         CLEO_SkipUnusedVarArgs(thread);
+
+        // set condition result
+        if (!IsLegacyScript(thread))
+        {
+            if (!returnArg) result &= 0xFF; // clip to AL
+            OPCODE_CONDITION_RESULT(result != 0);
+        }
+
         return OR_CONTINUE;
     }