+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - new opcode **2084 ([emulate_key_release](https://library.sannybuilder.com/#/sa/script/extensions/input/2084))**
 - new opcode **2085 ([get_controller_key](https://library.sannybuilder.com/#/sa/script/extensions/input/2085))**
 - new opcode **2086 ([get_key_name](https://library.sannybuilder.com/#/sa/script/extensions/input/2086))**
+- added warning message to commands **0AC2-0AC5** if used with non-3d audio streams
 
 ## 5.0.3
 - added **GxtHook.cleo** plugin to ignored list